+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog Natrium
 
+## v5.3.1 (02-11-2017)
+- Fixed a bug when app icon images are not 1024x1024
+
 ## v5.3.0 (26-10-2017)
 - Added unit tests
 - natrium.log output